pyproject.toml
[project]
name = "road-lane-segmentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "wandb>=0.24.0",
]
